refactor(deps): relocate the doctor subgraph to internal/doctor (decouple PR-B' slice D2)
The largest and final slice of relocating the monitor/supervisor's package-main
deps to internal/. Moves the entire doctor preflight subgraph out of cmd/bintrail
into a new internal/doctor package, so the supervisor's Doctor path (and the
future cmd/bintrail-console binary) reaches it without the cmd layer.

Moved cmd/bintrail/{doctor.go (machinery), doctor_capacity.go, diskfree_unix.go,
diskfree_other.go} → internal/doctor/{doctor.go, capacity.go, diskfree_*.go}.

Exported (externally referenced): Report (was doctorReport — up.go/monitor.go/
tests), CheckResult, CheckStatus, StatusPass/Fail/Warn/Skip, Build (was
buildDoctorReport), CapacityCheckName. Methods Write/Err/ErrExcluding were
already exported. KEPT PRIVATE (grep-checked: no external caller): the add
method, all check* funcs, connectWithoutDB, queryErrorRemediation,
isUnknownDatabaseErr, extractGrantUser, binlogRetentionMinSeconds, and every
capacity helper + diskFree.

cmd/bintrail/doctor.go shrinks 880→95 lines: keeps only the cobra command
(doctorCmd/init/runDoctor/parseDocRetain) + runDoctorTo as a thin wrapper over
doctor.Build + r.Write + r.Err. up.go's preflight and monitorSupervisor.Doctor
call doctor.Build; the supervisor maps *doctor.Report → *console.DoctorReport
(unchanged field copy, string(c.Status)). checkReplicationGrants calls
metadata.HasReplPrivileges (D1). internal/doctor imports cliutil/config/metadata/
indexer/mysql — cycle-free (none import doctor).

Tests follow their code: doctor_test.go + doctor_integration_test.go + the
capacity machinery tests + the capacity integration test → internal/doctor
(package doctor, can use the private add + the test-only sqlDB alias). Two tests
STAY in cmd because they exercise cmd funcs: TestParseDocRetain (parseDocRetain)
and TestUpPreflightOutcome (upPreflightOutcome) — the latter rebuilt to construct
doctor.Report via its exported Checks + count fields (Err reads r.Failed,
ErrExcluding reads r.Checks) rather than the now-unreachable private add(), so
Add stays unexported (no over-export). setupPartitionedTable (a shared cmd
integration-test helper) is duplicated as a small test-local helper in
internal/doctor rather than churning testutil + 7 cmd files — flagged for a
possible dedup-to-testutil follow-up.

Pure relocation (verbatim bodies) + the runDoctorTo/upPreflightOutcome rewires;
zero behavior change. After this, the supervisor's entire Doctor path lives in
internal/ → the PR-D pivot (move supervisor + up --console to cmd/bintrail-console)
is unblocked.

Verified: go build/vet clean (unit + integration); full unit suite green;
internal/doctor integration (check/capacity/index tests) + cmd doctor + monitor/
supervisor integration (doctor.Build via Doctor) + the full E2E pipeline all pass
against MySQL.
